python-samples/gcp-resource-manager/requirements.txt
google-cloud-resource-manager==1.10.1